Assignment 1, Part 2 Dependency Parser Feature Selection
ITERATION 6:
* Back-tracked to ITERATION 3, with the following changes:
	* Fixed preconditions in the transition code, preventing endless
	* looping on some input sentences
	* Updated feature to include number of parents AND number of
	* children for the current stack/buffer items

RESULTS:
(With both parents and children)
* UAS: 0.6464
* LAS: 0.5392

(With only parents):
* UAS: 0.6647
* LAS: 0.5508

(With only children):
* UAS: 0.6582
* LAS: 0.5433

(Same as ITERATION 2, but w/ Transition):
* UAS: 0.6281
* LAS: 0.5305

COMMENTS:
It would appear that the preconditions for transitions improved the
baseline performance when compared to ITERATION number 2.  Interestingly,
there seems to be little impact on whether # of children or # of parents
is included.